move SearchActivity
