+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
             android:name=".presentation.ui.library.LibraryActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.ui.tracks.SearchActivity"
+            android:name=".data.SearchActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.ui.settings.SettingsActivity"
