+++ app/src/main/java/com/example/playlistmaker/data/SearchActivity.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.presentation.ui.tracks
+package com.example.playlistmaker.data
 
 import android.content.Context
 import android.content.Intent
@@ -25,6 +25,7 @@ import com.example.playlistmaker.domain.api.TracksData
 import com.example.playlistmaker.domain.models.Track
 import com.example.playlistmaker.domain.models.TrackConsumerImpl
 import com.example.playlistmaker.presentation.ui.audioplayer.AudioPlayerActivity
+import com.example.playlistmaker.presentation.ui.tracks.TrackAdapter
 import com.google.gson.Gson
 
 class SearchActivity : AppCompatActivity() {
